backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.2,<5.0
PyMySQL>=1.1,<2.0
cryptography
python-dotenv>=1.0,<2.0
Pillow>=9.0,<11.0
